add new alchemy enahancements
