+++ src/scripts/potions.ts
@@ -23,15 +23,6 @@ namespace Scripts {
             return kad;
         }
 
-        static getMortar(): string {
-            const mortars = Orion.FindType(gameObject.uncategorized.mortar.graphic);
-            if (!mortars.length) {
-                Scripts.Utils.log(`Nemas mortar`, ColorEnum.red);
-                throw 'Nemas mortar';
-            }
-            return mortars[0];
-        }
-
         /**
          * @param potionName nazev potionu, ktery je definovan jako klic v objectu o.potions
          * @param switchWarModeWhenNeeded prepne war, pokud je potreba docepnout
@@ -186,107 +177,6 @@ namespace Scripts {
                 Scripts.Utils.log('Nenalezena drink hlaska v journalu za posledni vterinu', ColorEnum.red);
             }
             Orion.ClearJournal(`${succMsg}|${failMsg}|${paraMsg}`);
-        }
-
-        static gmMortar(potionName: PotionsEnum) {
-            if (!isPotionsEnum(potionName)) {
-                return;
-            }
-
-            const p = gameObject.potions[potionName];
-            const cilovaKadSerial = Scripts.Potions.getKadForPotion(p);
-
-            const isEmptyKad = Orion.FindType(
-                gameObject.uncategorized.emptyKad.graphic,
-                gameObject.uncategorized.emptyKad.color,
-            );
-            if (!isEmptyKad) {
-                Scripts.Utils.log('Nemas praznou kad', ColorEnum.red);
-                return;
-            }
-
-            Scripts.Utils.playerPrint(`Target gmmortar for making "${potionName}"`);
-            Orion.WaitForAddObject('gmMortar', 60000);
-
-            while (true) {
-                Orion.ClearJournal();
-                Orion.Wait(50);
-
-                const kadePrevious = Orion.FindType(gameObject.uncategorized.emptyKad.graphic);
-
-                Scripts.Utils.selectMenu('Vyber typ potionu', [p.gmMortarSelection]);
-                Orion.UseObject('gmMortar');
-                Scripts.Utils.waitWhileSomethingInJournal(['You put the Nadoba', 'Musis mit']);
-                if (Orion.InJournal('Musis mit')) {
-                    Scripts.Utils.log('Dosly regy', ColorEnum.red);
-                    return;
-                }
-
-                const kadeNew = Orion.FindType(gameObject.uncategorized.emptyKad.graphic);
-                const michnutaKadSerial = kadeNew.filter((i) => kadePrevious.indexOf(i) === -1)[0];
-
-                Orion.ClearJournal();
-                Orion.Wait(50);
-                Orion.WaitTargetObject(cilovaKadSerial);
-                Orion.UseObject(michnutaKadSerial);
-                Scripts.Utils.waitWhileSomethingInJournal(['Prelil jsi']);
-
-                Orion.ClearJournal();
-                Orion.Wait(responseDelay);
-                const emptyBottle = Scripts.Potions.getEmptyBottle();
-                Orion.WaitTargetObject(emptyBottle);
-                Orion.UseObject(michnutaKadSerial);
-                Scripts.Utils.waitWhileSomethingInJournal(['You put']);
-
-                Orion.ClearJournal();
-                Orion.Wait(50);
-                Orion.WaitTargetType(p.graphic, p.color);
-                Orion.UseObject(cilovaKadSerial);
-                Scripts.Utils.waitWhileSomethingInJournal(['You put']);
-            }
-        }
-
-        static alchemy(potionName: PotionsEnum) {
-            if (!isPotionsEnum(potionName)) {
-                return;
-            }
-
-            const p = gameObject.potions[potionName];
-            const mortar = Scripts.Potions.getMortar();
-
-            while (true) {
-                Scripts.Utils.worldSaveCheckWait();
-                Orion.ClearJournal();
-
-                Scripts.Utils.selectMenu('Vyber typ potionu', [p.alchemySelection]);
-                Orion.UseObject(mortar);
-
-                Scripts.Utils.waitWhileSomethingInJournal(['You completed', 'You toss', 'Nemas dostatecny'], 60000);
-
-                if (Orion.InJournal('Nemas dostatecny')) {
-                    Scripts.Utils.log('Dosly regy', ColorEnum.red);
-                    return;
-                }
-
-                if (Orion.InJournal('You toss')) {
-                    continue;
-                }
-
-                Scripts.Utils.worldSaveCheckWait();
-                Orion.ClearJournal();
-                Orion.UseObject(mortar);
-                Scripts.Utils.waitWhileSomethingInJournal(['You pour'], 60000);
-                Orion.Wait(responseDelay);
-
-                Scripts.Utils.worldSaveCheckWait();
-                Orion.ClearJournal();
-                const potion = Scripts.Utils.findFirstType(p);
-                const kad = Scripts.Potions.getKadForPotion(p);
-                Orion.WaitTargetObject(potion);
-                Orion.UseObject(kad);
-                Scripts.Utils.waitWhileSomethingInJournal(['You put'], 60000);
-                Orion.Wait(responseDelay);
-            }
         }
 
         static fillPotion(
